requirements.txt
# ── Core framework ────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1

# ── AI SDKs ───────────────────────────────────────────────────
anthropic==0.34.0
openai==1.45.0
google-generativeai==0.8.0

# ── Database ──────────────────────────────────────────────────
asyncpg==0.29.0          # PostgreSQL async driver
redis[asyncio]==5.0.8    # Redis async

# ── Memory / Search ───────────────────────────────────────────
# pgvector==0.3.2        # Uncomment for semantic vector search

# ── Tools ─────────────────────────────────────────────────────
aiohttp==3.10.5          # HTTP calls (scraper, API caller)
aiofiles==24.1.0         # Async file I/O
beautifulsoup4==4.12.3   # HTML parsing
duckduckgo-search==6.2.13 # Free web search

# ── Scheduler ─────────────────────────────────────────────────
apscheduler==3.10.4

# ── Bots ──────────────────────────────────────────────────────
python-telegram-bot==21.5
twilio==9.3.0            # WhatsApp via Twilio

# ── Security / Auth ───────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── Utilities ─────────────────────────────────────────────────
pydantic==2.9.0
pydantic-settings==2.5.2
loguru==0.7.2
